Create test_singularity_metrics.m
Tests the various metrics by giving it a bunch of singular and
non-singular examples. Some from our physical examples, some
contrived singular examples.
There are two situations that I thought were singular but that
are reporting that they are not. That needs to be resolved.
They current raise assert failures.