Complete ellipsoid plotting
Now relies on a helper function, since linear and angular both work the same way.

Could be improved with ability to set a location of the center and then
plot the ellipse on an existing figure, where you have the robot drawn.

Includes a wrapper that puts the two ellipses next to each other.